Minor function to log out button added

+++ library-management-client/Views/SidebarView.axaml.cs
@@ -6,6 +6,8 @@ using Avalonia.Controls;
 using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml;
 using Microsoft.Extensions.DependencyInjection;
+using MsBox.Avalonia;
+using MsBox.Avalonia.Enums;
 
 namespace Avalonia_DependencyInjection.Views;
 
@@ -21,10 +23,19 @@ public partial class SidebarView : UserControl
         Console.WriteLine("Hello");
     }
 
-    private void Button_OnClick(object? sender, RoutedEventArgs e)
+    private async void Button_OnClick(object? sender, RoutedEventArgs e)
     {
-        File.Delete("userToken.txt");
-        var win=App.AppHost.Services.GetRequiredService<MainWindowViewModel>();
-        win.ContentViewModel = App.AppHost.Services.GetRequiredService<LoginViewModel>();
+        var box = MessageBoxManager
+            .GetMessageBoxStandard("Logging out", "Do you want to log out?",
+                ButtonEnum.YesNo, Icon.Question);
+
+        var result = await box.ShowAsync();
+
+        if (result == ButtonResult.Yes)
+        {
+            File.Delete("userToken.txt");
+            var win=App.AppHost.Services.GetRequiredService<MainWindowViewModel>();
+            win.ContentViewModel = App.AppHost.Services.GetRequiredService<LoginViewModel>();
+        }
     }
 }
